Apply a round of clippy fixes from nightly

+++ src/bfs.rs
@@ -27,7 +27,7 @@ impl BFS {
         // initialize bit vector storing which nodes are targets
         temp.target_set.resize(number_of_nodes, false);
         for i in target_list {
-            temp.target_set.set(*i as usize, true);
+            temp.target_set.set(*i, true);
         }
 
         temp.populate_sources(number_of_nodes);
